Switch to Monaco
While this does mean we no longer have vim and emacs support, it does
allow for a nicer editing experience and more flexibility in the future.

To be honest, this is what I wanted to go with initially, but it was a
bit of a pain to build locally. We're going via a CDN, so this is much
easier.